- working on #50 - trying to make this robust against differences in naming.  better to get fields directly from the image header, and replace and spaces with an underscore. - skipping weight and bpm files - only remove junkfile2 if it exists

+++ python3/uat_sortobjects.py
@@ -205,8 +205,8 @@ for f in filters:
                 destination = os.path.join(objname,filename)
                 os.rename(filename, destination)
                 
-                
-os.remove('junkfile2')
+if os.path.exists('junkfile2'):            
+    os.remove('junkfile2')
